#25-add-user-state: move classes manipulating users in `user` package
